rename glt model for SenciNact
