+++ backend/de.jena.glt.sensinact.model/model/glt.ecore
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ecore:EPackage xmi:version="2.0" xmlns:xmi="http://www.omg.org/XMI" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore" name="glt" nsURI="https://model.data.jena.de/sensinact/glt/1.0" nsPrefix="glt">
+    xmlns:ecore="http://www.eclipse.org/emf/2002/Ecore" name="gltSensiNact" nsURI="https://model.data.jena.de/sensinact/glt/1.0"
+    nsPrefix="glt">
   <eAnnotations source="Version">
     <details key="value" value="1.0"/>
   </eAnnotations>
