+++ backend/de.jena.glt.sensinact.model/src/de/jena/glt/sensinact/model/gltSensiNact/GltFactory.java
@@ -11,7 +11,7 @@
  * Contributors:
  *     Data In Motion - initial API and implementation
  */
-package de.jena.glt.sensinact.model.glt;
+package de.jena.glt.sensinact.model.gltSensiNact;
 
 import org.eclipse.emf.ecore.EFactory;
 
@@ -22,7 +22,7 @@ import org.osgi.annotation.versioning.ProviderType;
  * The <b>Factory</b> for the model.
  * It provides a create method for each non-abstract class of the model.
  * <!-- end-user-doc -->
- * @see de.jena.glt.sensinact.model.glt.GltPackage
+ * @see de.jena.glt.sensinact.model.gltSensiNact.GltPackage
  * @generated
  */
 @ProviderType
@@ -33,7 +33,7 @@ public interface GltFactory extends EFactory {
 	 * <!-- end-user-doc -->
 	 * @generated
 	 */
-	GltFactory eINSTANCE = de.jena.glt.sensinact.model.glt.impl.GltFactoryImpl.init();
+	GltFactory eINSTANCE = de.jena.glt.sensinact.model.gltSensiNact.impl.GltFactoryImpl.init();
 
 	/**
 	 * Returns a new object of class '<em>Side</em>'.
