+++ backend/de.jena.glt.sensinact.model/src/de/jena/glt/sensinact/model/gltSensiNact/util/package-info.java
@@ -13,4 +13,4 @@
  */
 @org.osgi.annotation.bundle.Export
 @org.osgi.annotation.versioning.Version("1.0")
-package de.jena.glt.sensinact.model.glt.impl;
+package de.jena.glt.sensinact.model.gltSensiNact.util;
